breaking apart graph helper into smaller files.
